@@ -20,14 +20,14 @@
 		}
 	],
 	"require": {
-		"php": ">=7.4",
-		"cakephp/cakephp": "^4.2.0",
-		"dereuromark/cakephp-tools": "^2.0.0",
-		"dereuromark/cakephp-tinyauth": "^3.0.0"
+		"php": ">=8.1",
+		"cakephp/cakephp": "5.x-dev",
+		"dereuromark/cakephp-tools": "dev-cake5",
+		"dereuromark/cakephp-tinyauth": "dev-cake5"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^9.5",
-		"fig-r/psr2r-sniffer": "dev-master"
+		"phpunit/phpunit": "^9.5.16",
+		"fig-r/psr2r-sniffer": "dev-next"
 	},
 	"support": {
 		"source": "https://github.com/dereuromark/cakephp-tinyauth-backend"
@@ -58,5 +58,6 @@
 		"allow-plugins": {
 			"dealerdirect/phpcodesniffer-composer-installer": true
 		}
-	}
+	},
+	"minimum-stability": "dev"
 }
