@@ -8,8 +8,8 @@ use Tools\Model\Entity\Entity;
  * @property int $id
  * @property int $type
  * @property string $path
- * @property \Cake\I18n\FrozenTime|null $created
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $created
+ * @property \Cake\I18n\DateTime|null $modified
  */
 class AllowRule extends Entity {
 
@@ -32,7 +32,7 @@ class AllowRule extends Entity {
 	 *
 	 * @var array<string, bool>
 	 */
-	protected $_accessible = [
+	protected array $_accessible = [
 		'type' => true,
 		'path' => true,
 		'created' => true,
