@@ -12,7 +12,7 @@ class DbAllowAdapterTest extends TestCase {
 	 *
 	 * @var array
 	 */
-	protected $fixtures = [
+	protected array $fixtures = [
 		'plugin.TinyAuthBackend.TinyAuthAllowRules',
 	];
 
