@@ -21,7 +21,7 @@ class AllowControllerTest extends TestCase {
 	 *
 	 * @var array
 	 */
-	protected $fixtures = [
+	protected array $fixtures = [
 		'plugin.TinyAuthBackend.TinyAuthAllowRules',
 	];
 
