@@ -15,11 +15,11 @@ class Plugin extends BasePlugin {
 	 * @return void
 	 */
 	public function routes(RouteBuilder $routes): void {
-		$routes->prefix('Admin', function (RouteBuilder $routes) {
+		$routes->prefix('Admin', function (RouteBuilder $routes): void {
 			$routes->plugin(
 				'TinyAuthBackend',
 				['path' => '/auth'],
-				function (RouteBuilder $routes) {
+				function (RouteBuilder $routes): void {
 					$routes->connect('/', ['controller' => 'Auth', 'action' => 'index']);
 					$routes->fallbacks();
 				},
